Demo app tests: move getErrorMsg & makeFeatureEntry tests to separate files
